Document the public API, curate every DocC symbol, and fix the defects it exposed
Public-header documentation coverage goes from 70.7% to 96.7% (2,362 of 2,443
declarations). Every live public declaration now carries a HeaderDoc block; the
remainder are forward declarations and `extern "C"` braces. Public Swift API in
FxGripRealityKit reaches 100%.

DocC gains a curation layer. 723 top-level symbols sat in DocC's automatic
kind buckets, reachable only through an alphabetical dump: 477 macros, 116
variables, 47 functions, 29 enums, 25 protocols, 13 unions, 12 classes, 10
structs and 7 typealiases. All 723 are now curated into named topic groups on
the article that owns the concept, leaving the buckets empty. The parameter
configuration keys go to FxPlugParameters, the notification names to
ExtensionArchitecture, the per-control keys to each control's own article. Only
framework-wide symbols, the error codes and the version pair, sit on FxGrip.md.

Fourteen symbol extension files are added under FxGrip.docc/Extensions and
FxGripRealityKit.docc/Extensions. A file headed with a symbol link, such as
``FxGrip/FxGripTileableEffect-class``, attaches an authored overview and Topics
groups to that symbol's generated page. FxGripTileableEffect moves from three
alphabetical buckets to 31 named groups over 172 members. This is the per-class
documentation a separate catalog cannot provide, because a catalog binds to a
build target.

The audit exposed defects that the documentation had been masking.

- `lightingAPIv3_Raw` had no implementation: the method was named
  `lightingAPIv3_Raws`. The class-side property re-declaration auto-synthesized
  a getter over an ivar nothing assigns, so the accessor returned nil for every
  plug-in. Removing the re-declarations surfaced it.
- `pluginVersion` was `@dynamic` with no getter anywhere, so a call would have
  raised an unrecognized selector. It is now captured from the host at init
  behind a respondsToSelector: guard, as pluginUUID and sessionID already were.
- `FxGripAPIAccessing` held its state unsafely. `pluginUUID` was an `assign`
  NSString read from the host, now `copy`. `effect` was `assign` and dangled
  after the effect deallocated, now `weak` and nullable, matching
  FxGripExtensionBase. `apiAccessing` was `assign` on the host manager the
  wrapper depends on for its lifetime, now `strong`; the host manager holds no
  reference back, so no cycle forms.
- Fifteen doc blocks were silently detached. An ordinary `//` comment between a
  `/*!` block and its declaration severs the attachment, and the symbol renders
  with no abstract. FxGripTileableEffect, FxGripParameterBase,
  FxGripExtensionBase, four Custom* protocols and three FxGripStaticRegistrar
  methods were all losing their documentation this way.
- Six abstracts in FxGripDynamicParameterAPI_v4 described the wrong edge:
  `setParameter:floatSliderMaximum:` was documented as setting the minimum, as
  were all three integer slider variants.
- `objectAtIndexedSubscript:` and `objectForKeyedSubscript:` documented behavior
  the code does not have. The first is dual-mode, taking a parameter ID when
  positive and a negated ordinal position otherwise. The second returns an
  extension for a non-numeric string key, not a parameter by name.
- Three abstracts in the test suite had drifted away from their methods, leaving
  one test undocumented and mislabeling another.

Two bodies of dead code are removed. FxGripAPITransaction was 175 commented-out
lines in the header and implementation, never live in this repository's history
and referenced nowhere; FxGripOOBParameterAccess supersedes it with the same
scoped shape over a real FxPlug API. The FxGripAPIAccessing class re-declared 66
protocol properties that its own conformance already provides, doubling the
generated documentation for no gain; the four state properties that back ivars
stay.

Build changes. `-Wdocumentation-unknown-command` is enabled, with `-iframework`
entries marking the FxPlug SDK, /Library/Frameworks and the vendored Frameworks
directory as system headers, without which the flag adds 420 warnings from
third-party headers. The FxPlugStub target opts out, because it mirrors Apple's
headers verbatim. Sixteen warnings in framework code are fixed, covering
CoreGraphics enum mixing, a variable-length array, undeclared selectors, a
designated-initializer violation and a Swift UnsafeRawPointer. The host API
method dump that ran on every effect instantiation is now behind
FXGRIP_LOG_HOST_API_METHODS, off by default.

AGENTS.md records the diagnostic build flags, the enforced doc-tag rule, and the
indentation a markdown list needs inside an `@discussion` to render as a list
rather than a code block.

Full Check passes: build and tests on macOS arm64, tests on the x86_64 slice,
tests under AddressSanitizer, and docbuild. Zero documentation warnings across
the framework and the test targets, and zero DocC diagnostics.

+++ FxGrip/Parameters/Point/FxGripPointOptions.m
@@ -40,7 +40,7 @@
 		_rangeMinY = [self doubleFrom:config key:kFxGripPointKey_RangeMinY fallback:0.0];
 		_rangeMaxY = [self doubleFrom:config key:kFxGripPointKey_RangeMaxY fallback:1.0];
 
-		_coordinateMapping = [self doubleFrom:config key:kFxGripPointKey_CoordinateMapping fallback:0.0] == FxGripPointCoordinateQuartzComposer
+		_coordinateMapping = [self doubleFrom:config key:kFxGripPointKey_CoordinateMapping fallback:0.0] == (double)FxGripPointCoordinateQuartzComposer
 			? FxGripPointCoordinateQuartzComposer : FxGripPointCoordinatePixel;
 		_compensateFrameMargin = [self boolFrom:config key:kFxGripPointKey_CompensateFrameMargin fallback:NO];
